major refactor - split into smaller functions
